requirements.txt
backtrader
pandas
numpy
matplotlib
python-binance
scikit-learn
joblib